refactor: organize project structure and implement authentication context - Move config files to /config/, replace useAuth hooks with AuthContext, add documentation, fix API call loops
