+++ package.json
@@ -12,6 +12,7 @@
     "test:coverage": "jest --coverage",
     "prepare": "husky"
   },
+  "prettier": "./config/.prettierrc",
   "dependencies": {
     "@types/bcryptjs": "^2.4.6",
     "@types/jsonwebtoken": "^9.0.9",
