+++ src/app/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { useAuth } from "@/hooks/useAuth";
+import { useAuth } from "@/contexts/AuthContext";
 import Navigation from "@/components/Navigation";
 import CloudinaryTest from "@/components/CloudinaryTest";
 import { useEffect, useState } from "react";
